Learning Laravel
Upgraded to Laravel 5.6, added auth scaffold (using php artisan make:auth) and tweaked the blade templates so that they fit the basic layout I made.

Todo: when logged in, it takes you to the home view which has a dashboard section to the side of my dashboard area, so I need to combine them somehow.

Signed-off-by: Neil <neil.whitaker@gmail.com>